requirements.txt
Flask==2.3.3
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
pytesseract==0.3.10
playwright==1.39.0
pyautogui==0.9.54
pynput==1.7.6
psutil==5.9.6
requests==2.31.0
python-socketio==5.9.0
flask-socketio==5.3.6
schedule==1.2.0
pyyaml==6.0.1
loguru==0.7.2
watchdog==3.0.0
pywin32==306; sys_platform == "win32"